# config/NMDAOpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API CORONA FINANCIERO - ADMISION",
    "description": "API Corona Financiero Admision",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [{
    "url": "http://localhost:8080/admision/api",
    "description": "Generated server url"
  }],
...
  "paths": {
...
    "/informacion-oferta/rechazarOferta": {
      "post": {
        "tags": ["Fase Informacion Oferta"],
        "operationId": "rechazarOferta",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfertaRechazadaDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/informacion-oferta/aceptarOferta": {
      "post": {
        "tags": ["Fase Informacion Oferta"],
        "operationId": "aceptarOferta",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfertaAceptadaDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/cliente/datos/": {
      "post": {
        "tags": ["Cliente"],
        "operationId": "registroDatosCliente",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroClienteRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/biometria/verificacion": {
      "post": {
        "tags": ["Biometria"],
        "operationId": "verificaValidacionBiometrica",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivarDispositivoRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VerificacionBiometriaResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/biometria/prospecto-solicitud/{rut}/{solicitudId}": {
      "post": {
        "tags": ["Biometria"],
        "operationId": "registrarBiometriaProspecto",
        "parameters": [{
          "name": "rut",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }, {
          "name": "solicitudId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BiometriaRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BiometriaResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/biometria/excepcion/{rut}/{solicitudId}": {
      "post": {
        "tags": ["Biometria"],
        "operationId": "registrarBiometriaExcepcion",
        "parameters": [{
          "name": "rut",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }, {
          "name": "solicitudId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BiometriaRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BiometriaResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/biometria/activar": {
      "post": {
        "tags": ["Biometria"],
        "operationId": "activarDispositivo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivarDispositivoRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ActivarDispositivoResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/biometria/activar/excepcion": {
      "post": {
        "tags": ["Biometria"],
        "operationId": "activarDispositivoExcepcion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivarDispositivoRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ActivarDispositivoResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/perfilamiento/rol/{rolId}": {
      "get": {
        "tags": ["Perfilamiento"],
        "operationId": "obtenerRol",
        "parameters": [{
          "name": "rolId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RolDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Perfilamiento"],
        "operationId": "eliminarRol",
        "parameters": [{
          "name": "rolId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/informacion-oferta/evaluacionProducto/solicitud/{solicitudId}/rut/{rut}": {
      "get": {
        "tags": ["Fase Informacion Oferta"],
        "operationId": "listarProductosOfercidosBySolicitudIdAndRut",
        "parameters": [{
          "name": "solicitudId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64"
          }
        }, {
          "name": "rut",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductoCupoOfrecidoDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OfertaRechazadaDTO": {
        "type": "object",
        "properties": {
          "solicitudId": {
            "type": "integer",
            "format": "int64"
          },
          "rut": {
            "type": "string"
          },
          "idMotivoRechazo": {
            "type": "integer",
            "format": "int64"
          },
          "otroMotivo": {
            "type": "string"
          }
        }
      },
      "OfertaAceptadaDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          },
          "evaluacionProductoId": {
            "type": "integer",
            "format": "int64"
          },
          "diaPagoId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "RegistroClienteRequestDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          },
          "telefonoAdicional": {
            "type": "string"
          },
          "nacionalidad": {
            "type": "integer",
            "format": "int64"
          },
          "estadoCivil": {
            "type": "integer",
            "format": "int64"
          },
          "profesion": {
            "type": "integer",
            "format": "int64"
          },
          "actividad": {
            "type": "integer",
            "format": "int64"
          },
          "codigoRegion": {
            "type": "string"
          },
          "codigoComuna": {
            "type": "string"
          },
          "calle": {
            "type": "string"
          },
          "numero": {
            "type": "string"
          },
          "departamento": {
            "type": "string"
          },
          "referencias": {
            "type": "string"
          },
          "latitud": {
            "type": "number",
            "format": "float"
          },
          "longitud": {
            "type": "number",
            "format": "float"
          },
          "envioEcc": {
            "type": "boolean"
          },
          "envioSMS": {
            "type": "boolean"
          }
        }
      },
      "ActivarDispositivoRequestDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          },
          "solicitudAdmisionId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "VerificacionBiometriaResponseDTO": {
        "type": "object",
        "properties": {
          "validacionBiometrica": {
            "type": "boolean"
          },
          "validacionExcepcion": {
            "type": "boolean"
          },
          "solicitudId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BiometriaDataRequestDTO": {
        "type": "object",
        "properties": {
          "Cedula_Apellido": {
            "type": "string"
          },
          "Cedula_Rut": {
            "type": "string"
          },
          "Cedula_Serie": {
            "type": "string"
          },
          "Cedula_Tipo": {
            "type": "string"
          },
          "Cedula_Vencimiento": {
            "type": "string"
          },
          "CedulaHuellaCoinciden": {
            "type": "boolean"
          },
          "IDTransaccion": {
            "type": "string"
          }
        }
      },
      "BiometriaRequestDTO": {
        "type": "object",
        "properties": {
          "Resultado": {
            "type": "boolean"
          },
          "Codigo": {
            "type": "string"
          },
          "Detalle": {
            "type": "string"
          },
          "Data": {
            "$ref": "#/components/schemas/BiometriaDataRequestDTO"
          }
        }
      },
      "BiometriaResponseDTO": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "mensaje": {
            "type": "string"
          }
        }
      },
      "ActivarDispositivoResponseDTO": {
        "type": "object",
        "properties": {
          "rut": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "RolDTO": {
        "type": "object",
        "properties": {
          "rolId": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string"
          },
          "moduloId": {
            "type": "integer",
            "format": "int64"
          },
          "acciones": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "vigencia": {
            "type": "boolean"
          }
        }
      },
...
      "ProductoCupoOfrecidoDTO": {
        "type": "object",
        "properties": {
          "evaluacionProductoId": {
            "type": "integer",
            "format": "int64"
          },
          "descripcionTipoProducto": {
            "type": "string"
          },
          "cupoAprobado": {
            "type": "integer",
            "format": "int64"
          },
          "recomendado": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}